pyproject.toml
[project]
name = "market-simulator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.4.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "rich>=13.7.1",
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
    "dataclasses-json>=0.6.4",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "core",
    "engine",
    "market",
    "participants",
    "api",
    "user",
    "display",
]
